Dynamic sleep time based on EID pool size
The MCTP control daemon need to sleep based on the Endpoint ID
pool size.

This is recommended from the FPGA team to make sure the delay
provides enough time to generate the routing tables.

Also, change the bootup delay type to integer so that it can handle
more than 256 seconds.

+++ ctrld/mctp-ctrl-cmdline.h
@@ -25,7 +25,7 @@ typedef struct mctp_cmdline_args_ {
     char                    name[10];
     int                     device_id;
     bool                    verbose;
-    uint8_t                 delay;
+    int                     delay;
     mctp_binding_ids_t      binding_type;
     uint8_t                 bind_info[MCTP_PVT_BIND_BUFF_SIZE];
     int                     bind_len;
